requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
configparser==3.5.0
idna==2.6
numpy==1.14.3
opencv-python==3.4.0.12
Pillow==5.1.0
pytesseract==0.2.0
requests==2.18.4
selenium==3.11.0
urllib3==1.22
